+++ crates/pytauri-core/src/plugins/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "plugin-dialog")]
+mod dialog;
 #[cfg(feature = "plugin-notification")]
 mod notification;
 
@@ -15,7 +17,15 @@ pub mod pytauri_plugins {
     #[pymodule_export]
     pub const PLUGIN_NOTIFICATION: bool = cfg!(feature = "plugin-notification");
 
+    /// Whether the `plugin-dialog` feature is enabled.
+    #[pymodule_export]
+    pub const PLUGIN_DIALOG: bool = cfg!(feature = "plugin-dialog");
+
     #[cfg(feature = "plugin-notification")]
     #[pymodule_export]
     pub use notification::notification;
+
+    #[cfg(feature = "plugin-dialog")]
+    #[pymodule_export]
+    pub use dialog::dialog;
 }