+++ python/pytauri/src/pytauri_plugins/dialog/__init__.py
@@ -1,0 +1,27 @@
+"""[tauri_plugin_dialog::self](https://docs.rs/tauri-plugin-dialog/latest/tauri_plugin_dialog/index.html)"""
+
+from pytauri_plugins.dialog.ffi import (
+    DialogExt,
+    FileDialogBuilder,
+    FileDialogBuilderArgs,
+    FilePath,
+    ImplDialogExt,
+    MessageDialogBuilder,
+    MessageDialogBuilderArgs,
+    MessageDialogButtons,
+    MessageDialogButtonsType,
+    MessageDialogKind,
+)
+
+__all__ = [
+    "DialogExt",
+    "FileDialogBuilder",
+    "FileDialogBuilderArgs",
+    "FilePath",
+    "ImplDialogExt",
+    "MessageDialogBuilder",
+    "MessageDialogBuilderArgs",
+    "MessageDialogButtons",
+    "MessageDialogButtonsType",
+    "MessageDialogKind",
+]